+++ database/src/commonMain/kotlin/ru/pavlig43/database/NocombroDatabase.kt
@@ -11,7 +11,6 @@ import androidx.room.useWriterConnection
 import androidx.sqlite.driver.bundled.BundledSQLiteDriver
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
 import kotlinx.coroutines.launch
 import kotlinx.datetime.LocalDate
 import ru.pavlig43.core.TransactionExecutor
@@ -33,7 +32,7 @@ import ru.pavlig43.database.data.product.dao.ProductDeclarationDao
 import ru.pavlig43.database.data.transaction.Transaction
 import ru.pavlig43.database.data.transaction.TransactionProductBDIn
 import ru.pavlig43.database.data.transaction.dao.ProductTransactionDao
-import ru.pavlig43.database.data.transaction.reminder.Reminder
+import ru.pavlig43.database.data.transaction.reminder.ReminderBD
 import ru.pavlig43.database.data.transaction.reminder.dao.ReminderDao
 import ru.pavlig43.database.data.vendor.Vendor
 import ru.pavlig43.database.data.vendor.dao.VendorDao
@@ -57,7 +56,7 @@ import kotlin.time.ExperimentalTime
         Transaction::class,
         TransactionProductBDIn::class,
 
-        Reminder::class,
+        ReminderBD::class,
     ],
     version = 2,
     autoMigrations = [